scripts/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
pillow==9.2.0
    # via -r requirements.in
tqdm==4.54.0
    # via -r requirements.in
